avoid branded areas, tighten 'Done' button matches
Summary:
Two needles (at least, there may be more) fail to match properly
with Rawhide (F23) because the top part of some screens now has
branding, while before it was plain grey. On root_password_screen
we can just throw that match away, we don't really need it, the
text 'Root Password:' seems specific enough to be reliable. On
user_creation_screen let's replace the branded area match with a
couple of the text box labels.

This also tightens the 'Done' button match area on both spokes;
the original area is too loose and covers some of the now-branded
area outside the button, so the match fails.

This fixes the default_install test for 2015-06-07 nightly.

Test Plan:
Try a test run against the 2015-06-07 nightly and see if the
default_install test passes. Maybe also check that it still
passes for an F22 build.

Reviewers: garretraziel, jskladan

Reviewed By: jskladan

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D386