don't include radio button state in anaconda_minimal_highlighted
Summary:
There's no need to include the radio button state here, which
causes the match to fail if Minimal is highlighted and already
selected. Hitting Space in this case won't de-select it (it's
a radio button, not a checkbox), and will give the desired
result. So we should accept it.

Test Plan:
Run tests against Rawhide (and maybe F22 as a regression check)
and ensure the package set selection step works, in tests where
it's included.

Reviewers: garretraziel, jskladan

Reviewed By: jskladan

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D385