add versions of multiple needles for updated Cantarell
Cantarell 0.0.17, which landed in 23 Branched on 2015-10-09,
changes several elements of the font's appearance and leads to
many failed needle matches. This commit contains variants of
all appropriate needles.

We cannot simply replace all the needles as we are doing the
22 stable nightly tests as well. Once all tests we run are on
releases which have 0.0.17, we could make the cantarell17 files
into the 'standard' ones and remove the old ones.

Landing without review as this breaks all tests and the change
is straightforward, simply a case of updating the needles with
the interactive editor.