Cleaning up adamw's mess
Moved all the mess from repo root to a directory. Since there is no need
for all the needles to be in the repo root - no magic happening here, at
the deployment, the needles are copied in the needles directory anyway.

Landing without review, as this seems to be the trend now ;)