adjust openQA test suite names (install_foo not server_foo)
Summary:
We named a bunch of the	tests 'server_foo' back	when we	were
starting out and didn't	really know what we were doing.	They're
all really installation	tests, not 'server' tests. I actually
want to	start adding some Server tests now, so it seems	like a
good time to fix that mess. This standardizes on 'install_' as
a prefix for installation tests, converting all	the 'server_'
tests and fixing up a couple of	odd cases to use 'install_'.

Test Plan:
Apply along with the matching commit for tools, do
a full test run and report test, and make sure all tests run
with the new names and correct ResTups are generated for wiki
submission.

Reviewers: jskladan, garretraziel

Reviewed By: garretraziel

Subscribers: tflink

Differential Revision: https://phab.qadevel.cloud.fedoraproject.org/D790